+++ src/db/migrations/migrations.js
@@ -14,6 +14,7 @@ import m0009 from './0009_pantry.sql';
 import m0010 from './0010_list_pantry.sql';
 import m0011 from './0011_pantry_expiry.sql';
 import m0012 from './0012_products.sql';
+import m0013 from './0013_sets.sql';
 
   export default {
     journal,
@@ -30,7 +31,8 @@ m0008,
 m0009,
 m0010,
 m0011,
-m0012
+m0012,
+m0013
     }
   }